refactor(mcp): split tools god-module into definitions/parse/handlers (#277)
Behavior-preserving Extract-Module refactor of the 2552-line
`bin/mcp/src/tools/mod.rs` god module. Zero logic changes.

New layout under `tools/`:
- `definitions.rs` — `all_tool_definitions()` + `tool_def()` + the JSON
  schema blobs (the doc-test moves with it).
- `parse.rs` — argument parsing/validation/result-shaping helpers and the
  dump-path helpers, plus the full 37-test unit module (verbatim).
- `handlers/{p0,p1,p2,outcome,session,cognitive}.rs` — the 26 `handle_*`
  functions, grouped by the existing JSON-schema phase comments.
- `mod.rs` — now the dispatcher: declares the submodules, keeps the
  string-keyed `dispatch` match, and re-exports the public surface.

The public surface is unchanged: `tools::all_tool_definitions`,
`tools::dispatch`, `tools::MAX_FLUSH_INSIGHTS`, and
`tools::MAX_BOOTSTRAP_BYTES` resolve exactly as before. The `tests/*.rs`
integration suites compile and pass with no edits, proving path
preservation. Helper/handler fns are `pub` inside `pub(crate)` submodules
(the depth.rs precedent) so their effective visibility stays crate-local
without tripping `clippy::redundant_pub_crate`.

The 37-test count is preserved (now under `tools::parse::tests`); all unit,
integration, and doc tests pass and clippy is clean.

Closes #277. Part of #256.
